Shuffleboard Additions, Encoder Constant Changes
Added ability to change the output of forward/reverse, side-to-side and rotational movement individually on Shuffleboard.
Updated the encoder angle offsets again. May need to check to see if all magnets are locked in tight.
Changed AprilTag docking command to use 'field oriented' drive so that it takes advantage of the PID control using the gyro.